+++ views/characters.handlebars
@@ -1,0 +1,115 @@
+
+<h2>Characters</h2>
+
+<table id="table" border="1">
+   <thead>
+     <th>ID</th>
+     <th>First Name</th>
+     <th>Last Name</th>
+     <th>School</th>
+     <th>House</th>
+   </thead>
+   <tbody>
+      {{#each characters}}
+        <tr>
+          <td>{{characterId}}</td>
+          <td>{{fname}}</td>
+          <td>{{lname}}</td>
+          <td>{{schoolName}}</td>
+          <td>{{houseName}}</td>
+        </tr>
+      {{/each}}
+   </tbody>
+</table>
+
+<h1>Add Character</h1>
+<div>
+  <form id="addchar">
+     <label for="fname">First Name</label>
+     <input type="text" id="fname" name="fname" placeholder="Harry">
+
+     <label for="lname">Last Name</label>
+     <input type="text" id="lname" name="lname" placeholder="Potter">
+
+     <label for="school">School</label>
+     <select name="school">
+        {{#each characters}}
+         <option value="{{schoolId}}">{{schoolName}}</option>
+        {{/each}}
+     </select>
+
+     <label for="House">House</label>
+     <select name="house">
+         {{#each characters}}
+          <option value="{{id}}">{{fname}} {{lname}}</option>
+         {{/each}}
+     </select>
+  </form>
+
+  <input id="submit" type="submit" value="Submit" onclick="submit()">
+
+</div>
+
+<h1>Update Character</h1>
+<div>
+   <form id="form">
+     <label for="character">Character</label>
+     <select name="character">
+         {{#each characters}}
+          <option value="{{id}}">{{fname}} {{lname}}</option>
+         {{/each}}
+     </select>
+
+     <label for="fname">First Name</label>
+     <input type="text" id="fname" name="fname" placeholder="Harry">
+
+     <label for="lname">Last Name</label>
+     <input type="text" id="lname" name="lname" placeholder="Potter">
+
+     <label for="school">School</label>
+     <select name="school">
+        {{#each characters}}
+         <option value="{{id}}">{{schoolName}}</option>
+        {{/each}}
+     </select>
+
+     <label for="house">House</label>
+     <select name="house">
+        {{#each characters}}
+         <option value="{{id}}">{{houseName}}</option>
+        {{/each}}
+     </select>
+  </form>
+  <input id="submit" type="submit" value="Submit" onclick="updateCharacter()">
+</div>
+
+<h1>Delete Character</h1>
+<div>
+  <form id="form">
+     <label for="character">Character</label>
+       <select name="character">
+           {{#each characters}}
+            <option value="{{id}}">{{fname}} {{lname}}</option>
+           {{/each}}
+       </select>
+  </form>
+  <input id="submit" type="submit" value="Submit" onclick="deleteCharacter()">
+</div>
+
+<h1>Filter Characters</h1>
+
+  <form id="filter_chars">
+     <label for="school">School</label>
+     <select name="school">
+         {{#each characters}}
+          <option value="{{id}}">{{schoolName}}</option>
+         {{/each}}
+     </select>
+     <input type="button" value="Filter" onclick="filterCharacters()">
+  </form>
+
+  <form id="search_chars">
+     <label for="search">Search</label>
+     <input type="text" id="search" name="search" placeholder="Harry">
+     <input type="submit" value="Search" onclick="search()">   
+  </form>
