+++ views/layouts/main.handlebars
@@ -7,6 +7,7 @@
 <body>
 	<div>
 		<a href="/">Home</a>
+		<a href="/characters">Characters</a>
 		<a href="/schools">Schools</a>
 		<a href="/spells">Spells</a>
 	</div>
